requirements.txt
numpy
opencv-python
pillow
PyAutoGUI
pytesseract
pynput
requests
pyobjc
pywinctl
webhook